requirements.txt
streamlit
pandas
playwright
beautifulsoup4
python-dotenv
openai
